+++ src/error.rs
@@ -4,7 +4,6 @@ use std::fmt;
 #[derive(Debug)]
 pub enum Error {
     Message(String),
-    Zenkit(String),
 }
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
@@ -13,32 +12,36 @@ impl fmt::Display for Error {
             "{:?}",
             match self {
                 Error::Message(s) => s,
-                Error::Zenkit(s) => s,
             }
         )
     }
 }
 
+
+impl From<config::ConfigError> for Error {
+    fn from(e: config::ConfigError) -> Error {
+        Error::Message(format!("configuration: {}",e.to_string()))
+    }
+}
+
 impl From<handlebars::TemplateError> for Error {
     fn from(e: handlebars::TemplateError) -> Error {
-        println!("Template Error: {:#?}", e);
-        Error::Message(e.to_string())
+        Error::Message(format!("template: {}",e.to_string()))
     }
 }
 impl From<handlebars::RenderError> for Error {
     fn from(e: handlebars::RenderError) -> Error {
-        println!("Render Error: {:#?}", e);
-        Error::Message(e.to_string())
+        Error::Message(format!("template render: {}",e.to_string()))
     }
 }
 impl From<zenkit::Error> for Error {
     fn from(e: zenkit::Error) -> Error {
-        Error::Zenkit(e.to_string())
+        Error::Message(format!("zenkit: {}",e.to_string()))
     }
 }
 impl From<std::io::Error> for Error {
     fn from(e: std::io::Error) -> Error {
-        Error::Message(e.to_string())
+        Error::Message(format!("IO: {}",e.to_string()))
     }
 }